+++ poc/classes/AuxSTConstructor.py
@@ -1,5 +1,5 @@
 from poc.classes.AuxEvaluation import EvaluateParams
-from poc.classes.AuxInbuiltValues import InbuiltValueAtRuntime
+from poc.classes.AuxInbuiltValues import InbuiltValueAtRuntime, InbuiltValueUndefined
 from poc.classes.AuxSTBuildingBlock import AuxSTBuildingBlock
 from poc.classes.AuxSTConstants import AuxSTConstants
 from poc.classes.AuxSymbolTable import AuxSymbolTable
@@ -52,7 +52,7 @@ class AuxSTConstructor(AuxSTBuildingBlock):
             self.filter_misused_templates(error_mgr, filename)
 
     def evaluate(self, sem):
-        new_value = InbuiltValueAtRuntime(self, self.parent.parent)
+        new_value = InbuiltValueAtRuntime(self, self.get_declared_type())
         for child in self.children:
             if isinstance(child, AuxSTSignature):
                 EvaluateParams.evaluate_recursion(sem, child)
@@ -62,9 +62,12 @@ class AuxSTConstructor(AuxSTBuildingBlock):
                 # The semantics of any AuxSTIdentifier call inside the constructor of a class
                 # is to call the constructor of its base class. Therefore, we expect the type of the base class
                 ret = EvaluateParams.evaluate_recursion(sem, child, expected_type=self.get_base_class_type(sem))
-                new_value = ret.value
-                # todo: we still have to embed the variables and asserted predicates
-                # of the called base class inside the constructor's class instance
+                if ret.evaluation_error:
+                    new_value = InbuiltValueUndefined(self)
+                else:
+                    # todo: we still have to embed the variables and asserted predicates
+                    # of the called base class inside the constructor's class instance
+                    pass
             else:
                 raise NotImplementedError(str(type(child)))
         # the value of the constructor is a wrapper object with the type of its class